pyproject.toml
[project]
name = "claude-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.176.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.26.0",
    "itsdangerous>=2.2.0",
    "mcp[cli]>=1.11.0",
    "psycopg2>=2.9.10",
    "sqlalchemy>=2.0.41",
    "twilio>=9.6.5",
]
